Pass `item` directly to selection toggling
This was causing a bug in the checkbox event listener because when toggling the classList in that case, we were actually doing that for the checkbox itself instead of the `<li>` element wrapping it. This was causing the "selected" css styling to only trigger when clicking on the item only.