Add new view for creating folders
If a user does not find a folder they think may have existed, they have to exit the extension pop-up, go to the bookmark manager, find the parent folder and create their folder inside.

In order to avoid this extra work, the user can now directly search for the parent folder and directly create a subfolder underneath it without leaving the extension.